Address CodeRabbit PR #3 review: read-only guard, fingerprint, arg validation
- Reject non-read-only / multi-statement SQL on the simple-query path.
  telemetry::is_read_only_single parses and requires exactly one Statement::Query
  with no data-modifying CTE. The evaluator's parameterized GENERIC_PLAN branch
  (which uses raw_sql / the simple protocol, where `;`-separated statements would
  all execute) now refuses anything that isn't a single read-only query, and
  workload capture applies the same guard so a writable CTE
  (WITH … AS (DELETE …) SELECT …) or multi-statement text is never stored.
- Widen the fallback workload fingerprint from a truncated 24-bit hash to the
  full 64-bit digest (pgss-{:016x}) so distinct queries without a pg_stat_
  statements queryid can't collide and overwrite each other on upsert.
- `pistol capture` args are now u32 (limit >= 1), so negative values are
  rejected up front instead of erroring at the database.

Tests: unit coverage that the read-only guard rejects multi-statement text and
writable CTEs while accepting SELECT/read CTEs. Full suite green; verified live
that negatives are rejected and the self-driving capture→measure→apply path is
intact (+99.3% measured latency on a captured query).

+++ src/cli.rs
@@ -72,10 +72,10 @@ pub struct DemoAllArgs {
 pub struct CaptureArgs {
     /// Only capture queries called at least this many times.
     #[arg(long, default_value_t = 5)]
-    pub min_calls: i64,
+    pub min_calls: u32,
     /// Max number of (hottest) queries to capture.
-    #[arg(long, default_value_t = 50)]
-    pub limit: i64,
+    #[arg(long, default_value_t = 50, value_parser = clap::value_parser!(u32).range(1..))]
+    pub limit: u32,
 }
 
 #[derive(Args)]
@@ -139,7 +139,7 @@ pub async fn dispatch(cli: Cli) -> anyhow::Result<()> {
 }
 
 async fn capture(pool: &PgPool, args: CaptureArgs) -> anyhow::Result<()> {
-    let n = telemetry::capture_workload(pool, args.min_calls, args.limit).await?;
+    let n = telemetry::capture_workload(pool, args.min_calls as i64, args.limit as i64).await?;
     println!(
         "✓ captured {n} workload quer{} from pg_stat_statements (min_calls={}, limit={})",
         if n == 1 { "y" } else { "ies" },
